+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.
 
 ## [Unreleased]
 
+### Added
+
+- Support more entrypoints (and make finding it optional): `wmain`, `WinMain`, `wWinMain`
+
 ## [26] - 2026-08-29
 
 Initial release.